# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Oryon Gestao API - Documentacao de Endpoints",
        "description": "API RESTful profissional desenvolvida em Laravel 11 para gestao de departamentos, colaboradores e movimentacao estrategica de equipes.",
        "contact": {
            "name": "Jorge Carvalho",
            "email": "carvalhojorge7@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor Local de Desenvolvimento"
        }
    ],
    "paths": {
...
        "/api/employees": {
            "get": {
                "tags": [
                    "Colaboradores"
                ],
                "summary": "Listar colaboradores",
                "description": "Retorna listagem paginada de colaboradores com suporte a filtros por departamento e status.",
                "operationId": "b9b73e1873c5fe36b951f014f711e4fa",
                "parameters": [
                    {
                        "name": "department_id",
                        "in": "query",
                        "description": "Filtro por departamento",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtro por status (ativo ou inativo)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "ativo",
                                "inativo"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Registros por pagina",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numero da pagina",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listagem de colaboradores",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Nao autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Colaboradores"
                ],
                "summary": "Cadastrar colaborador",
                "description": "Cadastra um novo colaborador no sistema.",
                "operationId": "781d9d682a574c5e2c1dab17f782b5cd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "department_id",
                                    "role",
                                    "hired_at"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Carlos Eduardo Lima"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "carlos.lima@oryon.com.br"
                                    },
                                    "department_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "Desenvolvedor Backend Senior"
                                    },
                                    "hired_at": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-01-15"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "ativo",
                                        "enum": [
                                            "ativo",
                                            "inativo"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Colaborador cadastrado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validacao"
                    },
                    "401": {
                        "description": "Nao autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/employees/{id}": {
            "get": {
                "tags": [
                    "Colaboradores"
                ],
                "summary": "Exibir colaborador",
                "description": "Retorna os detalhes de um colaborador especifico.",
                "operationId": "de09c8d9afb3f7114d227ad0645940ac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do colaborador",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do colaborador"
                    },
                    "404": {
                        "description": "Colaborador nao encontrado"
                    },
                    "401": {
                        "description": "Nao autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Colaboradores"
                ],
                "summary": "Atualizar colaborador",
                "description": "Atualiza os dados cadastrais de um colaborador.",
                "operationId": "55e1ceec679a0d1436c869735e984a83",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do colaborador",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Carlos Eduardo Lima"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "Tech Lead"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "ativo",
                                        "enum": [
                                            "ativo",
                                            "inativo"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Colaborador atualizado com sucesso"
                    },
                    "404": {
                        "description": "Colaborador nao encontrado"
                    },
                    "422": {
                        "description": "Erro de validacao"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Colaboradores"
                ],
                "summary": "Inativar colaborador",
                "description": "Inativa o colaborador alterando seu status para inativo e aplicando Soft Delete.",
                "operationId": "dc1326c2f6b4bed0201d62396ddfd970",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do colaborador",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Colaborador inativado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Colaborador inativado com sucesso."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Colaborador nao encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Insira o token JWT retornado no endpoint /api/auth/login para autenticar as requisicoes.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}